feat: send 'add_to_cart' analytics event (#53)
* feat: send 'add_to_cart' analytics event

add logic to send 'add_to_cart' analytics event as a callback when buy button is pressed

* feat: create AnalyticsProvider component

* feat: add AnalyticsProvider to gatsby ssr and browser

* test: add cypress tests for add_to_cart analytics event

* fix: add second parameter on useEffect hook

Co-authored-by: Ícaro Azevedo <icaro.azevedo@vtex.com.br>

* fix: return children directly on AnalyticsProvider

instead of returning it wrapped into a react fragment

* fix: rename AnalyticsProvider to AnalyticsHandler

* docs: add todo comment
to explore other ways of calculating the product's value in the future

* refactor: change interface call to a cleaner one

instead of AddToCartEvent['data'] it is cleaner to use AddToCartData

* fix: adjust event to work with store-api

some graphql queries changed. Removed 'categoryTree' from queries since it is no longer available. Also, changed the query format for 'brand'.

* fix: move dataLayer creation to global scope

Co-authored-by: Tiago Gimenes <tlgimenes@gmail.com>

* fix: correct AnalyticsHandler merge

* test: add cypress tests to check relevant event fields

* fix: change fields passed to useBuyButton

* fix: regenerate generated graphql files

* fix: remove brand from add_to_cart event

* feat: Add brand and item name to analytics event

This was achieved by destructuring and mounting the item object again with only the properties allowed by store-api. The useBuyButton sends extra fields to analytics, and those fields need to be removed before calling the 'addItem' method that triggers the mutation that adds the item to the store-sdk cart.

* fix: remove sku_reference_id from analytics event item

Co-authored-by: Ícaro Azevedo <icaro.azevedo@vtex.com.br>
Co-authored-by: Tiago Gimenes <tlgimenes@gmail.com>