Fix Lighthouse's non-composited animation for the Add to Cart button
[Lighthouse was complaining about a non-composited animation][1] in the PDP's Add to Cart button.
After a lot of investigation, we found out the problem was within [the recently added `transition: background-color…`][2] and the very specific scenario of the button starting disabled and being quickly enabled (after the product's data was queried). This is a normal pattern. Except, because this change of state happened very early on the page load, it was captured by Lighthouse, and a background-color transition triggers a re-paint, a non-composite animation ([see the rendering pipeline][3])!

Considering we don't want to remove the CSS `transition: background-color…` because it's valid, and we cannot also change the behaviour of having the button start disabled till it's properly activated, my only solution was to replace the disabled state for [a SVG skeleton component][4], which seemed to fix the non-composited animation violation. Let us confirm when the CI runs the Lighthouse check.

- https://web.dev/non-composited-animations/

[1]: https://github.com/vtex-sites/base.store/runs/4653738587
[2]: https://github.com/vtex-sites/base.store/commit/509623a44afcfaf6f35a34d04e89d947dc97a89b#diff-2f5b8daef0d1fdc78e04593c1485d310a232ce29ae73ff66dd9ef9d7eb449d0dR11
[3]: https://web.dev/non-composited-animations/
[4]: https://skeletonreact.com/