Refactor button focus ring
This commit rolls affects the focus state of a button. It rolls back to the previous focus ring technique that makes use of `box-shadow` instead of a simple `outline` due to Safari's lack of support on rounded edges. On top of that, I added a new color token for the background color of the page's body, which is combined with the focus ring technique so it can mimic the `outline-offset` effect with perfection.

This commit also simplifies the tokens of the Theming POC page.